Fix Duplicate Records In Stats Parsing
Prerequisite for Enderal, otherwise the Bashed Patch fails to build for
the official DLC.

'criticalEffect': (bolt.Path(u'Skyrim.esm'), 0)

Critical effect is a FormID but was being parsed as an int.
Which meant the comparision for a change always gave a result of true.
When led to a high number of ITM when using the stats patcher.

Infernio:
 Turns out that this fixes a whole lot more than just ITMs - WB flat out
 errors when patching some mods with critical effects in them. I still
 have absolutely 0 clue how this patcher works, so I think it's best if
 we just disable critical effect patching for now.