cosaves: Some initial hierarchy refactoring
Simply plugging pluggy in (sorry) is not going to work, so refactoring
it is. Let's start with explicity marking what is and is not meant to
be abstract and/or internal.

Also takes the initial steps towards integrating pluggy by having it
extend ACoSaveFile, which allows us to decomission self.path (self.name
and self.tag were unused, so binned).

Note that I have absolutely no idea whether or not this still works.
Testing will happen once I've gotten a bit more of the refactoring done.