WIP: Set game-specific prefix for SLocalSavePath
Prerequisite for Enderal, which is the first game to use a prefix other
than 'Saves\' for this INI setting - otherwise, saves profiles won't
work with Enderal.

Touches a lot of code in many different places, so marked as WIP since
it will need lots of testing - but it seems to work as intended.